Extract CfgSampleClass fixtures to their own file (retire-emitter step 1)
The new-pipeline tests (Gaps, CompileBack, comparison-tree, …) all depend on
CfgSampleClass and its sibling fixtures, which lived in ControlFlowGraphTests.cs
alongside the old-stack ControlFlowGraph tests. Move the fixtures into a
dedicated CfgSampleClass.cs so the legacy-emitter deletion can remove the old
test file without taking the shared fixtures with it.

Pure move — the fixtures are fully name-qualified, so no usings are needed; the
old tests still reference them (same assembly). 516 decompiler tests green, zero
behavior change.
